refactor(blog): migrate blog data handling to new lib and remove mock data
- Replaced mock blog post data with dynamic loading from the new `lib/blog` module.
- Updated components to use `blogPosts` instead of `mockBlogPosts` for better data management.
- Removed outdated mock data files to streamline the codebase.
- Introduced utility functions for managing blog post data, enhancing maintainability.
